resolves #176 - Allow margin to be set per heading level
Not much of a Ruby guy, but we needed to modify asciidoctor-pdf to support `margin_top`, `margin_bottom`, and `line_height` to be set per heading level.  While it's entirely possible I'm missing an edge case, it seems fairly trivial/innocuous to pass the heading level into the `layout_heading` method.

Thanks for all of your work on this incredible project.

_Amended to correct bad copy/paste job resulting in application of
@theme[%(heading_h#{opts[:level]}_margin_top)] to both top_margin and bot_margin._