Add possibility for recto/verso background images (PR #923)
* Add possibility for recto/verso background images
  * page_background_image_recto (theme) or page-background-image-recto (document attribute)
  * page_background_image_verso (theme) or page-background-image-verso (document attribute)
* store page_bg_image as hash by side; add tests

This change allows the user to have alternating background images. The change is backwards compatible with the current "page_background_image". The user can either use the original or the newly proposed key. If the same background image is specified for both the recto and verso image, the result is the same as if the user had used "page_background_image". If the background image for a side is not specified, the "page_background_image" is used as a fallback.