Modify subset-fonts.pe to add floor/ceiling characters
This follows advice in

https://github.com/asciidoctor/asciidoctor-pdf/issues/1832

on how to add some missing characters in the M+ fallback font. I think
the floor/ceiling characters are the only ones not already present, but
my branch explicitly tags all the glyphs the Vulkan spec requires that
aren't in the bundled (non-fallback) fonts, for possible future
reference.

The change to the bundled font is very small - about 220 bytes - and I
hope it can be included with asciidoctor-pdf going forward.

I had problems building from the repository scripts, as I don't have
much of the infrastructure the entire repository expects. In addition
there was some sort of issue with a mis-named font in subset-fonts.sh
relative to HEAD at the time I did the change. Eventually I cobbled
together a shell script which modified subset-fonts.sh and let me build
the one font I needed (the M+ fallback) locally using fontforge, without
using podman, and have verified that works for us.